requirements.txt
discord.py
python-dotenv
discord-py-slash-command
discord-py-interactions